+++ index.d.ts
@@ -47,6 +47,7 @@ declare module 'plivo-browser-sdk/client' {
             useDefaultAudioDevice?: boolean;
             registrationRefreshTimer?: number;
             enableNoiseReduction?: boolean;
+            usePlivoStunServer?: boolean;
             reconnectOnHeartbeatFail?: boolean;
             dtmfOptions?: DtmfOptions;
     }
@@ -697,7 +698,10 @@ declare module 'plivo-browser-sdk/managers/callSession' {
             state: string;
             stirShakenState: string;
             extraHeaders: ExtraHeaders;
-            remoteCancelReason: string;
+            protocol: string;
+            originator: string;
+            reason: string;
+            errorCode: number;
     }
     export interface SignallingInfo {
             call_initiation_time?: number;
@@ -882,7 +886,7 @@ declare module 'plivo-browser-sdk/managers/callSession' {
             /**
                 * Get basic call information.
                 */
-            getCallInfo: (reason?: string) => CallInfo;
+            getCallInfo: (originator: string, protocol?: string, reason?: string, errorCode?: number) => CallInfo;
             /**
                 * Triggered when the user answers the call(Outgoing/Incoming) and got or received 200 OK.
                 * @param {Client} clientObject - client reference
@@ -942,7 +946,6 @@ declare module 'plivo-browser-sdk/managers/callSession' {
                 * @private
                 */
             constructor(options: CallSessionOptions);
-            extractReason(reasontag: string | undefined, reasonRegex: RegExp): string;
     }
 }